s1/requirements.txt
boto3==1.24.78
aws-cdk-lib==2.43.0
constructs==10.1.109
python-dotenv
